refactor(dashboard): extract useDebouncedValue from SearchPage to improve reusability
